added logic to check if bids have been placed on product for delete

+++ AuctionExpress.Service/ProductService.cs
@@ -52,14 +52,14 @@ namespace AuctionExpress.Service
                         ProductName = e.ProductName,
                         CategoryName = e.ProductCategoryCombo.CategoryName,
                         ProductQuantity = e.ProductQuantity,
-                      // ProductIsActive = e.DetermineIsActive(),
+                        // ProductIsActive = e.DetermineIsActive(),
                         ProductStartTime = e.ProductStartTime,
                         ProductCloseTime = e.ProductCloseTime,
                         MinimumSellingPrice = e.MinimumSellingPrice
                     }
                     );
                 return query.ToList();
-                
+
             }
         }
 
@@ -70,7 +70,7 @@ namespace AuctionExpress.Service
                 var query =
                     ctx
                     .Product
-                    .Where(e => e.ProductCloseTime > DateTimeOffset.Now && e.ProductStartTime<DateTimeOffset.Now)
+                    .Where(e => e.ProductCloseTime > DateTimeOffset.Now && e.ProductStartTime < DateTimeOffset.Now)
                     .Select(e => new ProductListItem
                     {
                         ProductId = e.ProductId,
@@ -94,7 +94,7 @@ namespace AuctionExpress.Service
                 var query =
                     ctx
                     .Product
-                    .Where(e => e.ProductCloseTime > DateTimeOffset.Now && e.ProductStartTime < DateTimeOffset.Now && e.ProductCategoryId ==Id)
+                    .Where(e => e.ProductCloseTime > DateTimeOffset.Now && e.ProductStartTime < DateTimeOffset.Now && e.ProductCategoryId == Id)
                     .Select(e => new ProductListItem
                     {
                         ProductId = e.ProductId,
@@ -126,7 +126,7 @@ namespace AuctionExpress.Service
                 return
                     new ProductDetail
                     {
-                        ProductId=entity.ProductId,
+                        ProductId = entity.ProductId,
                         ProductName = entity.ProductName,
                         ProductCategoryId = entity.ProductCategoryId,
                         ProductDescription = entity.ProductDescription,
@@ -163,7 +163,7 @@ namespace AuctionExpress.Service
             }
         }
 
-        public bool DeleteProduct(int productId)
+        public string DeleteProduct(int productId)
         {
             using (var ctx = new ApplicationDbContext())
             {
@@ -172,16 +172,24 @@ namespace AuctionExpress.Service
                         .Product
                         .Single(e => e.ProductId == productId);
 
+                if (entity.Seller.Id != _userId.ToString())
+                    return "UnAuthorized to make this change.";
+                if (entity.ProductBids.Count > 0)
+                    return "Product can not be deleted, as bids have been placed.  Please contact site administration for assistance.";
+
                 ctx.Product.Remove(entity);
 
-                return ctx.SaveChanges() == 1;
+                if (ctx.SaveChanges() == 1)
+                    return "Product successfully removed.";
+
+                return "";
             }
         }
 
         public string ValidateAuctionStatus(int id)
         {
             var prodDetail = GetProductById(id);
-            
+
             if (prodDetail == null)
                 return "Product has been removed or does not exist.";
             if (prodDetail.ProductSeller != _userId.ToString())
